added mytournamentspage, remade table into own component
