+++ src/components/Navigation/LoggedIn.js
@@ -3,10 +3,12 @@ import { useDispatch, useSelector } from "react-redux";
 import { logOut } from "../../store/user/actions";
 import { selectUser } from "../../store/user/selectors";
 import { Avatar, Popover, Position, Menu, Button, Pane } from "evergreen-ui";
+import { useHistory } from "react-router-dom";
 
 export default function LoggedIn() {
   const dispatch = useDispatch();
   const user = useSelector(selectUser);
+  const history = useHistory();
   return (
     <Pane
       border
@@ -22,11 +24,19 @@ export default function LoggedIn() {
       <Avatar src={user.avatar} marginRight="1rem" />
       <Popover
         position={Position.BOTTOM_LEFT}
-        content={
+        content={({ close }) => (
           <Menu>
             <Menu.Group>
               <Menu.Item icon="user">My Account</Menu.Item>
-              <Menu.Item icon="numbered-list">My Tournaments</Menu.Item>
+              <Menu.Item
+                icon="numbered-list"
+                onSelect={() => {
+                  history.push("/mytournaments");
+                  close();
+                }}
+              >
+                My Tournaments
+              </Menu.Item>
               <Menu.Item icon="calendar">Matches</Menu.Item>
             </Menu.Group>
             <Menu.Divider />
@@ -40,7 +50,7 @@ export default function LoggedIn() {
               </Menu.Item>
             </Menu.Group>
           </Menu>
-        }
+        )}
       >
         <Button width="100%">{user.userName}</Button>
       </Popover>
